+++ publishing/publish-nexus.sh
@@ -8,7 +8,7 @@ echo
 version=$(echo *.msi | cut -d'-' -f2)
 mvn_args=""
 mvn_idx=1
-for ext in msi pkg zip tar.gz; do
+for ext in msi pkg appimage zip tar.gz; do
   for file in *.$ext; do
     mvn_args="$mvn_args -F maven2.asset$mvn_idx=@$file -F maven2.asset$mvn_idx.extension=$ext -Fmaven2.asset$mvn_idx.classifier=$(basename $file .$ext | cut -d'-' -f3-)"
     ((mvn_idx+=1))